Refactor: Move filename generation logic to utils
