+++ img/favicons/favicon-animal.nyc-64px.ico.meta
@@ -1,7 +1,7 @@
 {
-  "originalUrl": "https://www.google.com/s2/favicons?domain=animal.nyc&sz=256",
-  "downloadedAt": "2025-10-05T16:27:02.616Z",
+  "originalUrl": "https://www.google.com/s2/favicons?domain=animal.nyc&sz=64",
+  "downloadedAt": "2025-10-05T16:35:50.899Z",
   "type": "favicon",
   "filename": "favicon-animal.nyc-64px.ico",
-  "size": "256"
+  "size": "64"
 }
