+++ img/favicons/favicon-chunk-party.com-64px.ico.meta
@@ -1,7 +1,7 @@
 {
-  "originalUrl": "https://www.google.com/s2/favicons?domain=www.chunk-party.com&sz=256",
-  "downloadedAt": "2025-10-05T16:27:02.276Z",
+  "originalUrl": "https://www.google.com/s2/favicons?domain=www.chunk-party.com&sz=64",
+  "downloadedAt": "2025-10-05T16:35:50.473Z",
   "type": "favicon",
   "filename": "favicon-chunk-party.com-64px.ico",
-  "size": "256"
+  "size": "64"
 }
