+++ img/favicons/favicon-dieselseattle.com-64px.ico.meta
@@ -1,7 +1,7 @@
 {
-  "originalUrl": "https://www.google.com/s2/favicons?domain=dieselseattle.com&sz=256",
-  "downloadedAt": "2025-10-05T16:27:02.691Z",
+  "originalUrl": "https://www.google.com/s2/favicons?domain=dieselseattle.com&sz=64",
+  "downloadedAt": "2025-10-05T16:35:50.988Z",
   "type": "favicon",
   "filename": "favicon-dieselseattle.com-64px.ico",
-  "size": "256"
+  "size": "64"
 }
