+++ img/favicons/favicon-eagle-ny.com-64px.ico.meta
@@ -1,7 +1,7 @@
 {
-  "originalUrl": "https://www.google.com/s2/favicons?domain=eagle-ny.com&sz=256",
-  "downloadedAt": "2025-10-05T16:27:02.440Z",
+  "originalUrl": "https://www.google.com/s2/favicons?domain=eagle-ny.com&sz=64",
+  "downloadedAt": "2025-10-05T16:35:50.682Z",
   "type": "favicon",
   "filename": "favicon-eagle-ny.com-64px.ico",
-  "size": "256"
+  "size": "64"
 }
