+++ img/favicons/favicon-eventbrite.com-64px.ico.meta
@@ -1,7 +1,7 @@
 {
-  "originalUrl": "https://www.google.com/s2/favicons?domain=www.eventbrite.com&sz=256",
-  "downloadedAt": "2025-10-05T16:27:01.973Z",
+  "originalUrl": "https://www.google.com/s2/favicons?domain=www.eventbrite.com&sz=64",
+  "downloadedAt": "2025-10-05T16:35:50.075Z",
   "type": "favicon",
   "filename": "favicon-eventbrite.com-64px.ico",
-  "size": "256"
+  "size": "64"
 }
