+++ img/favicons/favicon-furball.nyc-64px.ico.meta
@@ -1,7 +1,7 @@
 {
-  "originalUrl": "https://www.google.com/s2/favicons?domain=www.furball.nyc&sz=256",
-  "downloadedAt": "2025-10-05T16:27:02.200Z",
+  "originalUrl": "https://www.google.com/s2/favicons?domain=www.furball.nyc&sz=64",
+  "downloadedAt": "2025-10-05T16:35:50.387Z",
   "type": "favicon",
   "filename": "favicon-furball.nyc-64px.ico",
-  "size": "256"
+  "size": "64"
 }
