+++ img/favicons/favicon-jackhammerchicago.com-64px.ico.meta
@@ -1,7 +1,7 @@
 {
-  "originalUrl": "https://www.google.com/s2/favicons?domain=jackhammerchicago.com&sz=256",
-  "downloadedAt": "2025-10-05T16:27:02.355Z",
+  "originalUrl": "https://www.google.com/s2/favicons?domain=jackhammerchicago.com&sz=64",
+  "downloadedAt": "2025-10-05T16:35:50.599Z",
   "type": "favicon",
   "filename": "favicon-jackhammerchicago.com-64px.ico",
-  "size": "256"
+  "size": "64"
 }
