+++ img/favicons/favicon-northalsted.com-64px.ico.meta
@@ -1,7 +1,7 @@
 {
-  "originalUrl": "https://www.google.com/s2/favicons?domain=northalsted.com&sz=256",
-  "downloadedAt": "2025-10-05T16:27:02.129Z",
+  "originalUrl": "https://www.google.com/s2/favicons?domain=northalsted.com&sz=64",
+  "downloadedAt": "2025-10-05T16:35:50.303Z",
   "type": "favicon",
   "filename": "favicon-northalsted.com-64px.ico",
-  "size": "256"
+  "size": "64"
 }
