+++ tools/download-images.js
@@ -8,7 +8,7 @@ const { URL } = require('url');
 const { JSDOM } = require('jsdom');
 
 // Import shared filename utilities
-const { generateFilenameFromUrl, generateFaviconFilename, generateEventFilename, cleanImageUrl, convertImageUrlToLocalPath } = require('../js/filename-utils.js');
+const { generateFilenameFromUrl, generateFaviconFilename, generateEventFilename, generateFilename, generateLinktreeFaviconFilename, cleanImageUrl, convertImageUrlToLocalPath } = require('../js/filename-utils.js');
 
 // Mock logger for Node.js environment
 global.logger = {
@@ -317,45 +317,6 @@ function downloadFile(url, outputPath, timeout = 30000, maxRedirects = 5) {
   });
 }
 
-// Generate filename from URL using shared utility
-function generateFilename(url, type = 'event', size = null) {
-    if (type === 'favicon') {
-        const baseFilename = generateFaviconFilename(url);
-        if (size) {
-            // Check if filename already contains a size suffix to avoid double suffixes
-            const ext = path.extname(baseFilename);
-            const nameWithoutExt = path.basename(baseFilename, ext);
-            
-            // If the filename already contains a size suffix (like -64px), don't add another one
-            if (nameWithoutExt.includes('-64px') || nameWithoutExt.includes('-32px') || nameWithoutExt.includes('-256px')) {
-                return baseFilename;
-            }
-            
-            // Add size suffix for higher quality favicons
-            return `${nameWithoutExt}-${size}px${ext}`;
-        }
-        return baseFilename;
-    }
-    return generateFilenameFromUrl(url);
-}
-
-// Generate a unique filename for Linktree profile pictures based on the Linktree URL
-function generateLinktreeFaviconFilename(linktreeUrl, size = '32') {
-    try {
-        const parsedUrl = new URL(linktreeUrl);
-        const pathname = parsedUrl.pathname.substring(1); // Remove leading slash
-        const cleanPath = pathname
-            .replace(/[^a-zA-Z0-9._-]/g, '-') // Replace invalid chars with dashes
-            .replace(/-+/g, '-') // Collapse multiple dashes
-            .replace(/^-|-$/g, ''); // Remove leading/trailing dashes
-        
-        return `favicon-linktr.ee-${cleanPath}-${size}px.png`;
-    } catch (error) {
-        // Fallback to hash-based filename
-        const hash = simpleHash(linktreeUrl);
-        return `favicon-linktr.ee-${hash}-${size}px.png`;
-    }
-}
 
 // Download image with a custom filename
 async function downloadImageWithCustomFilename(imageUrl, customFilename, type = 'event', isLinktreeProfile = false, targetSize = 96) {
